pyproject.toml
[project]
name = "trove"
version = "0.0.0"
description = "Open-source parsers, agents, and visualizations for underused healthcare datasets."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "Chris Betz" }]

[tool.uv.workspace]
members = ["packages/*"]

[dependency-groups]
dev = [
    "pillow>=12.2.0",
    "pytest>=8",
    "ruff>=0.8",
]
docs = [
    "mkdocs-material>=9.5",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["packages"]
addopts = "--import-mode=importlib"
